refactor: split input parser
