+++ src/@types/input-parser.ts
@@ -1,0 +1,6 @@
+import type { FormattedComment } from "@/@types/format.formatted";
+
+export interface InputParser {
+  key: string[];
+  parse: (input: unknown) => FormattedComment[];
+}
