+++ src/input/empty.ts
@@ -1,0 +1,8 @@
+import type { FormattedComment, InputParser } from "@/@types";
+
+export const EmptyParser: InputParser = {
+  key: ["empty"],
+  parse: (): FormattedComment[] => {
+    return [];
+  },
+};
